moved utils to util
