+++ main.py
@@ -1,6 +1,6 @@
 import argparse
 
-from src.utils.parser import load_parameters
+from src.util.parser import load_parameters
 from controller import Controller
 
 def main():
